Adding reproduction spec for #378
